Bug fix: Function images did not show sometimes
There is a bug that sometimes adds "img/pictos/" to the imgCell field of cells that didn't have any image assigned. Due to this, function images of function cells did not show, as imgCell prevailed over the rest of the images. Could not find where "img/pictos/ is added, so I made an extra condition while checking for empty imgCell in cells.