Bug fix: Could not press menu buttons
When the screen was not wide enough and menu buttons overlapped, when the menu bar was hidden, you could not press the function buttons to Read, Erase, etc. You could only press the bottom part of the button as the rest was overlapped by the menu with a higher zIndex.